Added my third graph.
Included a brief explanation on my linear model, and modified it so that it was internally consistent with the reasoning.

Added an automatic model comparison, although much of the heavy lifting is done by prior code. This model comparison checks all three of our metrics to see which model is favored, and picks the model that's favored by the majority of the metrics. If all three metrics pick a different model, then the winner is whichever model is favored by the RMSE.